Feat - completed code to support multiple configs.
