+++ Cargo.toml
@@ -42,6 +42,7 @@ parking_lot = "0.12.3"
 textwrap = "0.16.2"
 encoding_rs = "0.8.35"
 encoding_rs_io = "0.1.7"
+ctor = "0.4.1"
 
 [build-dependencies]
 clap = { version = "4.5.31", features = ["derive", "cargo", "env"] }
