+++ src/config/mod.rs
@@ -1,4 +1,9 @@
+mod color_conversion;
 pub mod files;
+pub mod format;
+mod glob_generation;
 pub mod load;
 pub mod log;
-mod raw;
+pub mod site;
+mod site_raw;
+mod tool_raw;
