+++ src/search_engine/cache.rs
@@ -1,7 +1,7 @@
 use crate::cli::command::CacheMode;
 use crate::config::files::database_path;
-use crate::config::load::Config;
 use crate::errors::error::IsError;
+use crate::search_engine::link::HtmlSource;
 use once_cell::sync::Lazy;
 use parking_lot::Mutex;
 use parking_lot::MutexGuard;
@@ -14,7 +14,7 @@ use zstd::{decode_all, encode_all};
 static HTML_CACHE: Lazy<Cache> = Lazy::new(Cache::new);
 static VERSION: u16 = 0;
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Default)]
 pub struct CacheConfig {
     cache_mode: CacheMode,
     max_size: usize,
@@ -35,16 +35,13 @@ impl CacheConfig {
 
 struct Cache {
     connection: Arc<Mutex<Connection>>,
-    config: CacheConfig,
 }
 
 impl Cache {
     pub fn new() -> Self {
         let conn = Connection::open(database_path()).expect("Failed to open database");
-        let config = Config::get_cache_config();
         let cache = Cache {
             connection: Arc::new(Mutex::new(conn)),
-            config,
         };
         cache.init_db().expect("Failed to initialize database");
         cache
@@ -63,21 +60,23 @@ impl Cache {
         Ok(())
     }
 
-    pub fn insert(&self, key: &str, value: &str) -> Result<(), IsError> {
-        match self.config.cache_mode {
+    pub fn insert(&self, key: &HtmlSource, value: &str) -> Result<(), IsError> {
+        let binding = key.get_config();
+        let cache_config = binding.get_cache();
+        match cache_config.cache_mode {
             CacheMode::Read | CacheMode::Never => return Ok(()),
             _ => {}
         }
-        let timestamp = Self::current_time()? + self.config.ttl;
+        let timestamp = Self::current_time()? + cache_config.ttl;
         let cache_size = self.get_cache_size()?;
-        if cache_size >= self.config.max_size {
-            self.purge_cache()?;
+        if cache_size >= cache_config.max_size {
+            self.purge_cache(cache_config)?;
         }
         let compressed_html = encode_all(Cursor::new(value), 3)?;
         self.get_connection().execute(
             "INSERT INTO cache (url, html, timestamp, version) VALUES (?, ?, ?, ?)
              ON CONFLICT(url) DO UPDATE SET html = excluded.html, timestamp = excluded.timestamp",
-            params![key, compressed_html, timestamp, VERSION],
+            params![key.get_url(), compressed_html, timestamp, VERSION],
         )?;
         Ok(())
     }
@@ -90,9 +89,9 @@ impl Cache {
         Ok(count)
     }
 
-    fn purge_cache(&self) -> Result<(), IsError> {
+    fn purge_cache(&self, cache_config: &CacheConfig) -> Result<(), IsError> {
         let elements_to_retain =
-            (self.config.max_size * (100 - self.config.cull as usize) + 99) / 100;
+            (cache_config.max_size * (100 - cache_config.cull as usize) + 99) / 100;
         self.get_connection()
             .prepare(
                 "DELETE FROM cache WHERE url IN (
@@ -104,17 +103,17 @@ impl Cache {
         Ok(())
     }
 
-    pub fn get(&self, key: &str) -> Result<Option<String>, IsError> {
-        match self.config.cache_mode {
+    pub fn get(&self, key: &HtmlSource) -> Result<Option<String>, IsError> {
+        match key.get_config().get_cache().cache_mode {
             CacheMode::Write | CacheMode::Never => return Ok(None),
             _ => {}
         }
 
-        let values = self.retrieve_value(key)?;
+        let values = self.retrieve_value(key.get_url())?;
 
         if let Some((html, timestamp)) = values {
             if timestamp <= Self::current_time()? {
-                log::debug!("Expired cache item {key}");
+                log::debug!("Expired cache item {}", key.get_url());
                 self.remove(key)?;
                 return Ok(None);
             }
@@ -139,14 +138,15 @@ impl Cache {
         }
     }
 
-    pub fn remove(&self, key: &str) -> Result<(), IsError> {
-        if let CacheMode::Never | CacheMode::Read = self.config.cache_mode {
+    pub fn remove(&self, key: &HtmlSource) -> Result<(), IsError> {
+        if let CacheMode::Never | CacheMode::Read = key.get_config().get_cache().cache_mode {
             return Ok(());
         }
         let connection = self.get_connection();
-        connection.execute("DELETE FROM cache WHERE url = ?", params![key])?;
+        let url = key.get_url();
+        connection.execute("DELETE FROM cache WHERE url = ?", params![url])?;
         drop(connection);
-        log::debug!("Removing: {} from cache", key);
+        log::debug!("Removing: {url} from cache");
         Ok(())
     }
 
@@ -166,18 +166,19 @@ impl Cache {
     }
 }
 
-pub fn cached_pages_write(url: &str, html: &str) {
-    log::debug!("Caching {}", url);
+pub fn cached_pages_write(url: &HtmlSource, html: &str) {
+    log::debug!("Caching {}", url.get_url());
     HTML_CACHE
         .insert(url, html)
         .unwrap_or_else(|e| log::error!("Error when writing page to cache: {:?}", e));
 }
 
-pub fn cached_pages_read(url: &str) -> Option<String> {
-    let cache_result = HTML_CACHE.get(url).unwrap_or_else(|e| {
+pub fn cached_pages_read(html_source: &HtmlSource) -> Option<String> {
+    let cache_result = HTML_CACHE.get(html_source).unwrap_or_else(|e| {
         log::error!("Error when reading page from cache: {:?}", e);
         None
     });
+    let url = html_source.get_url();
     match cache_result {
         Some(_) => log::debug!("Cache hit for {url}"),
         None => log::debug!("Cache miss for {url}"),
@@ -185,7 +186,7 @@ pub fn cached_pages_read(url: &str) -> Option<String> {
     cache_result
 }
 
-pub fn cached_pages_purge(url: &str) {
+pub fn cached_pages_purge(url: &HtmlSource) {
     HTML_CACHE.remove(url).unwrap_or_else(|e| {
         log::error!("Error when removing page from cache: {:?}", e);
     });
@@ -200,6 +201,9 @@ pub fn clear() {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::search_engine::link::tests::TEST_CONFIG;
+    use crate::search_engine::link::HtmlSource::LinkSource;
+    use crate::search_engine::link::Link;
     use serial_test::serial;
     use std::env;
     use tempfile::TempDir;
@@ -213,29 +217,42 @@ mod tests {
         let path = temp_dir.path();
         env::set_var("XDG_DATA_HOME", path);
         let _ = Cache::new(); // Just to init db in temp dir
+        TEST_CONFIG.write().cache = CacheConfig::new(CacheMode::ReadWrite, 2, MS_IN_SECOND * 5, 50);
         let path = database_path();
         let cache = Connection::open(path).expect("Failed to open cache");
         let cache = Cache {
             connection: Arc::new(Mutex::new(cache)),
-            config: CacheConfig::new(CacheMode::ReadWrite, 2, MS_IN_SECOND * 5, 50),
         };
         for i in 1..=5 {
             cache
                 .insert(
-                    format!("http://localhost:8080/{i}").as_str(),
+                    &LinkSource(Link::new(format!("http://localhost:8080/{i}").as_str())),
                     format!("html{i}").as_str(),
                 )
                 .unwrap();
         }
-        assert!(cache.get("http://localhost:8080/1").unwrap().is_none());
-        assert!(cache.get("http://localhost:8080/2").unwrap().is_none());
-        assert!(cache.get("http://localhost:8080/3").unwrap().is_none());
+        assert!(cache
+            .get(&LinkSource(Link::new("http://localhost:8080/1")))
+            .unwrap()
+            .is_none());
+        assert!(cache
+            .get(&LinkSource(Link::new("http://localhost:8080/2")))
+            .unwrap()
+            .is_none());
+        assert!(cache
+            .get(&LinkSource(Link::new("http://localhost:8080/3")))
+            .unwrap()
+            .is_none());
         assert_eq!(
-            cache.get("http://localhost:8080/4").unwrap(),
+            cache
+                .get(&LinkSource(Link::new("http://localhost:8080/4")))
+                .unwrap(),
             Some(String::from("html4"))
         );
         assert_eq!(
-            cache.get("http://localhost:8080/5").unwrap(),
+            cache
+                .get(&LinkSource(Link::new("http://localhost:8080/5")))
+                .unwrap(),
             Some(String::from("html5"))
         );
     }
@@ -248,24 +265,39 @@ mod tests {
         env::set_var("XDG_DATA_HOME", path);
 
         let _ = Cache::new(); // Just to init db in temp dir
+        TEST_CONFIG.write().cache = CacheConfig::new(CacheMode::ReadWrite, 2, 0, 50);
         let path = database_path();
         let cache = Connection::open(path).expect("Failed to open cache");
         let cache = Cache {
             connection: Arc::new(Mutex::new(cache)),
-            config: CacheConfig::new(CacheMode::ReadWrite, 2, 0, 50),
         };
         for i in 1..=5 {
             cache
                 .insert(
-                    format!("http://localhost:8080/{i}").as_str(),
+                    &LinkSource(Link::new(format!("http://localhost:8080/{i}").as_str())),
                     format!("html{i}").as_str(),
                 )
                 .unwrap();
         }
-        assert!(cache.get("http://localhost:8080/1").unwrap().is_none());
-        assert!(cache.get("http://localhost:8080/2").unwrap().is_none());
-        assert!(cache.get("http://localhost:8080/3").unwrap().is_none());
-        assert!(cache.get("http://localhost:8080/4").unwrap().is_none());
-        assert!(cache.get("http://localhost:8080/5").unwrap().is_none());
+        assert!(cache
+            .get(&LinkSource(Link::new("http://localhost:8080/1")))
+            .unwrap()
+            .is_none());
+        assert!(cache
+            .get(&LinkSource(Link::new("http://localhost:8080/2")))
+            .unwrap()
+            .is_none());
+        assert!(cache
+            .get(&LinkSource(Link::new("http://localhost:8080/3")))
+            .unwrap()
+            .is_none());
+        assert!(cache
+            .get(&LinkSource(Link::new("http://localhost:8080/4")))
+            .unwrap()
+            .is_none());
+        assert!(cache
+            .get(&LinkSource(Link::new("http://localhost:8080/5")))
+            .unwrap()
+            .is_none());
     }
 }
